requirements.txt
streamlit
groq
PyPDF2
python-docx
beautifulsoup4
